+++ tests/test_route_validation.py
@@ -34,7 +34,7 @@ def get_re_view(username: str):
 
 
 @pytest.mark.parametrize("username", [('busthead'), ('123'), ('pd_az')])
-def test_str_route_validation_valid(create_router_event, username):
+def test_str_route_validation_valid(create_api_event, username):
     """
     The string based route validation will match any valid \w+ regular expression,
     which is used for unicode patterns [a-zA-Z0-9_].
@@ -42,7 +42,7 @@ def test_str_route_validation_valid(create_router_event, username):
     https://docs.python.org/3/library/re.html
     """
 
-    event = create_router_event('/bio/{username}',
+    event = create_api_event('/bio/{username}',
                                 method='GET',
                                 pathParameters={'username': username})
 
@@ -53,14 +53,14 @@ def test_str_route_validation_valid(create_router_event, username):
 
 
 @pytest.mark.parametrize("username", [('$$$'), ('12#3'), ('hello@gmail')])
-def test_str_route_validation_invalid(create_router_event, username):
+def test_str_route_validation_invalid(create_api_event, username):
     """
     Test string based routing with INVALID paramters. Note that partial matches
     will be rejected. For instance hello@gmail will be rejected given that `@`
     is not part of [a-zA-Z0-9_] pattern.
     """
 
-    event = create_router_event('/bio/{username}',
+    event = create_api_event('/bio/{username}',
                                 method='GET',
                                 pathParameters={'username': username})
 
@@ -75,9 +75,9 @@ def get_articles_view(year: int, month: int):
     return {'year': year, 'month': month}
 
 
-def test_route_with_int_valid(create_router_event):
+def test_route_with_int_valid(create_api_event):
 
-    event = create_router_event('/articles/{year}/{month}/',
+    event = create_api_event('/articles/{year}/{month}/',
                                 method='GET',
                                 pathParameters={'year': '2020', 'month': '10'})
 
@@ -90,9 +90,9 @@ def test_route_with_int_valid(create_router_event):
 @pytest.mark.parametrize("year,month", [
     ('2020', 'INVALID'), ('INVALID', '12'), ('hello', 'world'),
 ])
-def test_route_with_int_invalid_params(create_router_event, year, month):
+def test_route_with_int_invalid_params(create_api_event, year, month):
 
-    event = create_router_event('/articles/{year}/{month}/',
+    event = create_api_event('/articles/{year}/{month}/',
                                 method='GET',
                                 pathParameters={'year': year, 'month': month})
 
@@ -106,13 +106,13 @@ def get_product(product_id: uuid.UUID):
     return {'id': str(product_id)}
 
 
-def test_uuid_in_route_valid(create_router_event):
+def test_uuid_in_route_valid(create_api_event):
     """
     Validate a uuid based parameter with a valid value.
     """
 
     pid = '00010203-0405-0607-0809-0a0b0c0d0e0f'
-    event = create_router_event('/product/{product_id}/',
+    event = create_api_event('/product/{product_id}/',
                                 method='GET',
                                 pathParameters={'product_id': pid})
 
@@ -125,12 +125,12 @@ def test_uuid_in_route_valid(create_router_event):
     ('00010203-0405-0607-0809'),
     ('INVALID'),
 ])
-def test_uuid_in_route_invalid(create_router_event, product_id):
+def test_uuid_in_route_invalid(create_api_event, product_id):
     """
     Validate the uiid based view with invalid values.
     """
 
-    event = create_router_event('/product/{product_id}/',
+    event = create_api_event('/product/{product_id}/',
                                 method='GET',
                                 pathParameters={'product_id': product_id})
 
@@ -147,12 +147,12 @@ def get_item(item_id: ReStr(r'([0-9a-f]{8}$)')):
 @pytest.mark.parametrize("item_id", [
     ('00010203'), ('52342512'), ('00102c03')
 ])
-def test_custom_re_in_route_valid(create_router_event, item_id):
+def test_custom_re_in_route_valid(create_api_event, item_id):
     """
     Validate a uuid based parameter with a valid value.
     """
 
-    event = create_router_event('/item/{item_id}/',
+    event = create_api_event('/item/{item_id}/',
                                 method='GET',
                                 pathParameters={'item_id': item_id})
 
@@ -164,12 +164,12 @@ def test_custom_re_in_route_valid(create_router_event, item_id):
 @pytest.mark.parametrize("item_id", [
     ('04052523209'), ('060809'), ('#0010203'), ('00102i03')
 ])
-def test_custom_re_in_route_invalid(create_router_event, item_id):
+def test_custom_re_in_route_invalid(create_api_event, item_id):
     """
     Validate the uiid based view with invalid values.
     """
 
-    event = create_router_event('/item/{item_id}/',
+    event = create_api_event('/item/{item_id}/',
                                 method='GET',
                                 pathParameters={'item_id': item_id})
 
@@ -192,13 +192,13 @@ def get_tracker_info(name: RouteTracker):
 @pytest.mark.parametrize("tracker_name", [
     ('fitbit'), ('nikeplus'), ('vivosmart')
 ])
-def test_custom_field_in_route_valid(create_router_event, tracker_name):
+def test_custom_field_in_route_valid(create_api_event, tracker_name):
     """
     Validate that a route with a custom field definition works when valid values
     are provided.
     """
 
-    event = create_router_event('/tracker/{name}/',
+    event = create_api_event('/tracker/{name}/',
                                 method='GET',
                                 pathParameters={'name': tracker_name})
 
@@ -210,12 +210,12 @@ def test_custom_field_in_route_valid(create_router_event, tracker_name):
 @pytest.mark.parametrize("tracker_name", [
     ('not there'), ('other')
 ])
-def test_custom_field_in_route_invalid(create_router_event, tracker_name):
+def test_custom_field_in_route_invalid(create_api_event, tracker_name):
     """
     Custom route field with invalid values.
     """
 
-    event = create_router_event('/tracker/{name}/',
+    event = create_api_event('/tracker/{name}/',
                                 method='GET',
                                 pathParameters={'name': tracker_name})